몬스터  - 테이블 데이터 몬스터 스탯들 적용하기

+++ Assets/Scripts/Scene/MainScene.cs
@@ -190,7 +190,7 @@ public class MainScene : MonoBehaviour
         return foundComponent;
     }
 
-    public void GetGoods(int getGold = 0, int getGem = 0)
+    public void GetGoods(double getGold = 0, double getGem = 0)
     {
         if (getGold > 0)
         {
